Separate Outbreak inner classes
